Refactoring/bug-fixing of T_StartTestOpportunity

+++ student/src/main/java/tds/student/performance/dao/mappers/TestOpportunityMapper.java
@@ -22,7 +22,7 @@ public class TestOpportunityMapper implements RowMapper<TestOpportunity> {
         testOpportunity.setSessionKey(UuidAdapter.getUUIDFromBytes(resultSet.getBytes("sessionKey")));
         testOpportunity.setBrowserKey(UuidAdapter.getUUIDFromBytes(resultSet.getBytes("browserKey")));
         testOpportunity.setTestKey(resultSet.getString("testKey"));
-        testOpportunity.setTestee(resultSet.getDouble("testee"));
+        testOpportunity.setTestee(resultSet.getLong("testee"));
         testOpportunity.setTestId(resultSet.getString("testId"));
         testOpportunity.setOpportunity(resultSet.getInt("opportunity"));
         testOpportunity.setStatus(resultSet.getString("status"));
